Moving all factories to OfflineModule
